Move sync-attribution tests into main test file
